restore sidebar settings footer
