+++ docs/chat-chain-changes/2026-06-16-sidebar-settings-footer.md
@@ -1,0 +1,8 @@
+---
+date: 2026-06-16
+pr: 1585
+feature: Sidebar settings footer
+impact: Removes the settings popover from chat page sidebars and restores profile/model/status controls to the app sidebar footer without changing chat session chain or message persistence behavior.
+---
+
+The chat, history, and group chat page sidebar Settings buttons now navigate directly to the settings page. Profile, model, status, language, theme, version, logout, and update controls live in the app sidebar footer used by settings and other non-page-sidebar routes.
