+++ packages/client/src/components/layout/AppSidebar.vue
@@ -1,13 +1,21 @@
 <script setup lang="ts">
-import { computed } from "vue";
+import { computed, ref } from "vue";
 import { useRoute, useRouter } from "vue-router";
 import { useI18n } from "vue-i18n";
+import { NButton, NModal, useMessage } from "naive-ui";
 import { useAppStore } from "@/stores/hermes/app";
 import { usePersistentRecord } from '@/composables/usePersistentRecord'
 import RouteLinkItem from '@/components/common/RouteLinkItem.vue'
-import { isStoredSuperAdmin } from "@/api/client";
+import ModelSelector from "@/components/layout/ModelSelector.vue";
+import ProfileSelector from "@/components/layout/ProfileSelector.vue";
+import LanguageSwitch from "@/components/layout/LanguageSwitch.vue";
+import ThemeSwitch from "@/components/layout/ThemeSwitch.vue";
+import VersionManagementModal from "@/components/layout/VersionManagementModal.vue";
+import { changelog } from "@/data/changelog";
+import { getStoredUsername, isStoredSuperAdmin } from "@/api/client";
 
 const { t } = useI18n();
+const message = useMessage();
 const route = useRoute();
 const router = useRouter();
 const appStore = useAppStore();
@@ -15,7 +23,13 @@ const selectedKey = computed(() => {
   return route.name as string;
 });
 const isSuperAdmin = computed(() => isStoredSuperAdmin());
+const currentUsername = computed(() => getStoredUsername());
 const isVersionPreview = import.meta.env.VITE_HERMES_PREVIEW === '1';
+const isDesktopShell = computed(() =>
+  (window as typeof window & { hermesDesktop?: { isDesktop?: boolean } }).hermesDesktop?.isDesktop === true,
+);
+const showChangelog = ref(false);
+const showVersionManagement = ref(false);
 
 function hasRoute(name: string): boolean {
   return router.hasRoute(name);
@@ -48,26 +62,36 @@ function handleSidebarClick(event: MouseEvent) {
     appStore.closeSidebar();
   }
 }
+
+async function handleUpdate() {
+  const ok = await appStore.doUpdate();
+  if (ok) {
+    message.success(t('sidebar.updateSuccess'), { duration: 5000 });
+  } else {
+    message.error(t('sidebar.updateFailed'));
+  }
+}
+
+function handleReloadClient() {
+  appStore.reloadClient();
+}
+
+function handleLogout() {
+  localStorage.clear();
+  window.location.reload();
+}
+
+function openChangelog() {
+  showChangelog.value = true;
+}
+
+function openVersionManagement() {
+  showVersionManagement.value = true;
+}
 </script>
 
 <template>
   <aside class="sidebar" :class="{ open: appStore.sidebarOpen, collapsed: appStore.sidebarCollapsed }" @click="handleSidebarClick">
-    <div class="sidebar-top-actions">
-      <RouteLinkItem class="nav-item sidebar-return-tab" :to="{ name: 'hermes.chat' }" :title="t('sidebar.backToChat')">
-        <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="1.7" stroke-linecap="round" stroke-linejoin="round">
-          <polyline points="15 18 9 12 15 6" />
-          <line x1="9" y1="12" x2="21" y2="12" />
-        </svg>
-        <span>{{ t("sidebar.backToChat") }}</span>
-      </RouteLinkItem>
-      <button class="collapse-btn" @click="appStore.toggleSidebarCollapsed()" :title="appStore.sidebarCollapsed ? t('sidebar.expand') : t('sidebar.collapse')">
-        <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-          <polyline v-if="appStore.sidebarCollapsed" points="9 18 15 12 9 6" />
-          <polyline v-else points="15 18 9 12 15 6" />
-        </svg>
-      </button>
-    </div>
-
     <nav class="sidebar-nav">
       <!-- Agent -->
       <div class="nav-group">
@@ -260,6 +284,99 @@ function handleSidebarClick(event: MouseEvent) {
         </div>
       </div>
     </nav>
+
+    <ProfileSelector />
+    <ModelSelector />
+
+    <div class="sidebar-footer">
+      <button class="nav-item logout-item" @click="handleLogout">
+        <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round">
+          <path d="M9 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h4" />
+          <polyline points="16 17 21 12 16 7" />
+          <line x1="21" y1="12" x2="9" y2="12" />
+        </svg>
+        <span>{{ t("sidebar.logout") }}</span>
+        <span v-if="currentUsername" class="logout-username" :title="currentUsername">{{ currentUsername }}</span>
+      </button>
+      <div class="status-row">
+        <div
+          class="status-indicator"
+          :class="{
+            connected: appStore.connected,
+            disconnected: !appStore.connected,
+          }"
+        >
+          <span class="status-dot"></span>
+          <span class="status-text">{{
+            appStore.connected
+              ? t("sidebar.connected")
+              : t("sidebar.disconnected")
+          }}</span>
+        </div>
+        <LanguageSwitch />
+      </div>
+      <div class="version-info">
+        <div class="version-links">
+          <a class="sidebar-footer-link" href="https://github.com/EKKOLearnAI/hermes-studio" target="_blank" rel="noopener noreferrer" title="GitHub">
+            <svg width="14" height="14" viewBox="0 0 24 24" fill="currentColor"><path d="M12 0C5.37 0 0 5.37 0 12c0 5.31 3.435 9.795 8.205 11.385.6.105.825-.255.825-.57 0-.285-.015-1.23-.015-2.235-3.015.555-3.795-.735-4.035-1.41-.135-.345-.72-1.41-1.23-1.695-.42-.225-1.02-.78-.015-.795.945-.015 1.62.87 1.845 1.23 1.08 1.815 2.805 1.305 3.495.99.105-.78.42-1.305.765-1.605-2.67-.3-5.46-1.335-5.46-5.925 0-1.305.465-2.385 1.23-3.225-.12-.3-.54-1.53.12-3.18 0 0 1.005-.315 3.3 1.23.96-.27 1.98-.405 3-.405s2.04.135 3 .405c2.295-1.56 3.3-1.23 3.3-1.23.66 1.65.24 2.88.12 3.18.765.84 1.23 1.905 1.23 3.225 0 4.605-2.805 5.625-5.475 5.925.435.375.81 1.095.81 2.22 0 1.605-.015 2.895-.015 3.3 0 .315.225.69.825.57A12.02 12.02 0 0024 12c0-6.63-5.37-12-12-12z"/></svg>
+          </a>
+          <a class="sidebar-footer-link" href="https://hermes-studio.ai/" target="_blank" rel="noopener noreferrer" title="Website">
+            <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><circle cx="12" cy="12" r="10"/><line x1="2" y1="12" x2="22" y2="12"/><path d="M12 2a15.3 15.3 0 0 1 4 10 15.3 15.3 0 0 1-4 10 15.3 15.3 0 0 1-4-10 15.3 15.3 0 0 1 4-10z"/></svg>
+          </a>
+        </div>
+        <span
+          class="version-text"
+          role="button"
+          tabindex="0"
+          @click="openChangelog"
+          @keydown.enter="openChangelog"
+          @keydown.space.prevent="openChangelog"
+        >
+          Studio v{{ appStore.serverVersion || "0.1.0" }}
+        </span>
+        <ThemeSwitch />
+      </div>
+      <NButton v-if="isDesktopShell" type="primary" size="tiny" block class="update-btn" @click="openVersionManagement">
+        {{ t('sidebar.versionManagement') }}
+      </NButton>
+      <NButton v-if="appStore.clientOutdated" type="warning" size="tiny" block class="update-btn" @click="handleReloadClient">
+        {{ t('sidebar.reloadClientVersion', { version: appStore.serverVersion }) }}
+      </NButton>
+      <NButton v-if="appStore.updateAvailable" type="primary" size="tiny" block class="update-btn" :loading="appStore.updating" @click="handleUpdate">
+        {{ appStore.updating ? t('sidebar.updating') : t('sidebar.updateVersion', { version: appStore.latestVersion }) }}
+      </NButton>
+    </div>
+
+    <div class="sidebar-top-actions">
+      <RouteLinkItem class="nav-item sidebar-return-tab" :to="{ name: 'hermes.chat' }" :title="t('sidebar.backToChat')">
+        <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="1.7" stroke-linecap="round" stroke-linejoin="round">
+          <polyline points="15 18 9 12 15 6" />
+          <line x1="9" y1="12" x2="21" y2="12" />
+        </svg>
+        <span>{{ t("sidebar.backToChat") }}</span>
+      </RouteLinkItem>
+      <button class="collapse-btn" @click="appStore.toggleSidebarCollapsed()" :title="appStore.sidebarCollapsed ? t('sidebar.expand') : t('sidebar.collapse')">
+        <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
+          <polyline v-if="appStore.sidebarCollapsed" points="9 18 15 12 9 6" />
+          <polyline v-else points="15 18 9 12 15 6" />
+        </svg>
+      </button>
+    </div>
+
+    <NModal v-model:show="showChangelog" preset="dialog" :title="t('sidebar.changelog')" style="width: 520px;">
+      <div class="changelog-list">
+        <div v-for="entry in changelog" :key="entry.version" class="changelog-version-block">
+          <div class="changelog-version-header">
+            <span class="changelog-version-tag">v{{ entry.version }}</span>
+            <span class="changelog-date">{{ entry.date }}</span>
+          </div>
+          <ul class="changelog-changes">
+            <li v-for="(change, idx) in entry.changes" :key="idx">{{ t(change) }}</li>
+          </ul>
+        </div>
+      </div>
+    </NModal>
+    <VersionManagementModal v-if="isDesktopShell" v-model:show="showVersionManagement" />
   </aside>
 </template>
 
@@ -383,7 +500,9 @@ function handleSidebarClick(event: MouseEvent) {
   display: flex;
   align-items: center;
   gap: 6px;
-  margin-bottom: 4px;
+  margin-top: 8px;
+  padding-top: 8px;
+  border-top: 1px solid $border-color;
 }
 
 .sidebar-return-tab {
@@ -391,6 +510,186 @@ function handleSidebarClick(event: MouseEvent) {
   min-width: 0;
   padding: 8px 10px;
   font-size: 13px;
+}
+
+.sidebar-footer {
+  padding-top: 10px;
+  border-top: 1px solid $border-color;
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.logout-item {
+  color: $text-secondary;
+
+  &:hover {
+    color: $error;
+  }
+
+  > span:not(.logout-username) {
+    min-width: 0;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+  }
+}
+
+.logout-username {
+  margin-left: auto;
+  max-width: 96px;
+  color: $text-muted;
+  font-size: 12px;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.status-row {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 8px;
+  padding: 2px 0 4px;
+}
+
+.status-indicator {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  min-width: 0;
+  padding-left: 12px;
+  font-size: 12px;
+  color: $text-secondary;
+
+  &.connected .status-dot {
+    background-color: $success;
+    box-shadow: 0 0 6px rgba(var(--success-rgb), 0.5);
+  }
+
+  &.disconnected .status-dot {
+    background-color: $error;
+  }
+}
+
+.status-dot {
+  width: 8px;
+  height: 8px;
+  border-radius: 50%;
+  flex-shrink: 0;
+}
+
+.status-text {
+  min-width: 0;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.version-info {
+  padding: 2px 0 8px 12px;
+  font-size: 11px;
+  color: $text-muted;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 6px;
+  overflow: hidden;
+}
+
+.version-links {
+  display: flex;
+  align-items: center;
+  flex-shrink: 0;
+  gap: 6px;
+}
+
+.sidebar-footer-link {
+  color: $text-muted;
+  display: flex;
+  align-items: center;
+  transition: color $transition-fast;
+
+  &:hover {
+    color: $text-primary;
+  }
+}
+
+.version-text {
+  flex: 0 0 auto;
+  overflow: visible;
+  white-space: nowrap;
+  cursor: pointer;
+  transition: color $transition-fast;
+
+  &:hover {
+    color: $accent-primary;
+  }
+}
+
+.version-info :deep(.theme-switch-container) {
+  flex-shrink: 0;
+}
+
+.update-btn {
+  margin: 4px 0 0;
+  border-radius: $radius-sm;
+}
+
+.changelog-list {
+  max-height: min(70vh, 640px);
+  overflow-y: auto;
+}
+
+.changelog-version-block {
+  margin-bottom: 20px;
+
+  &:last-child {
+    margin-bottom: 0;
+  }
+}
+
+.changelog-version-header {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  margin-bottom: 8px;
+}
+
+.changelog-version-tag {
+  font-weight: 600;
+  font-size: 14px;
+  color: $text-primary;
+  font-family: $font-code;
+}
+
+.changelog-date {
+  font-size: 12px;
+  color: $text-muted;
+}
+
+.changelog-changes {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+
+  li {
+    font-size: 13px;
+    color: $text-secondary;
+    padding: 4px 0 4px 16px;
+    position: relative;
+
+    &::before {
+      content: '';
+      position: absolute;
+      left: 0;
+      top: 12px;
+      width: 6px;
+      height: 6px;
+      border-radius: 50%;
+      background: $text-muted;
+    }
+  }
 }
 
 // ─── Collapsed sidebar (icon-rail mode) ─────────────────────────
@@ -408,7 +707,8 @@ function handleSidebarClick(event: MouseEvent) {
   .sidebar-top-actions {
     flex-direction: column;
     gap: 6px;
-    margin-bottom: 8px;
+    margin-top: 8px;
+    padding-top: 8px;
   }
 
   .sidebar-return-tab {
@@ -444,6 +744,28 @@ function handleSidebarClick(event: MouseEvent) {
     svg {
       flex-shrink: 0;
     }
+  }
+
+  :deep(.model-selector) {
+    display: none;
+  }
+
+  :deep(.profile-selector) {
+    display: flex;
+    justify-content: center;
+    padding: 8px 0;
+  }
+
+  .sidebar-footer {
+    align-items: center;
+    gap: 6px;
+    padding-top: 8px;
+  }
+
+  .status-row,
+  .version-info,
+  .update-btn {
+    display: none;
   }
 
 }
